Fix some main.sh errors, especially $my_shell
The biggest issue is that $my_shell was sometimes coming out
as `/bin/bash` rather than `bash`. For this reason and others,
I modified the `fzf` variable sourcing to only happen when a
wildcard match for `*bash` or `*zsh` actually work properly;
in other cases (which could happen if the shell is particularly
weirdly named, or if I try out e.g. fish) I do nothing